git-svn-id: http://mewsplayground-svn.cvsdude.com/seaweeds/trunk/bsponmpi@182 09f15e90-c184-4a26-aad6-528d2e8cadf7

+++ include/bsp_cpp/bsp_commandline.h
@@ -49,7 +49,7 @@ namespace bsp {
 	) {
 
 		if (p_from >= 0) {
-			int argc = args.size();
+			size_t argc = args.size();
 			bsp_broadcast(0, argc);
 
 			if (bsp_pid() != p_from) {
